fastapi-backend/requirements.txt
aiofiles==24.1.0
aiohttp==3.9.3
aioredis==2.0.1
aiosignal==1.3.2
aiosmtplib==3.0.2
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
billiard==4.2.1
black==24.10.0
boto3==1.34.149
botocore==1.34.149
cachetools==5.5.2
celery==5.4.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.8.2
cryptography==45.0.2
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
environs==14.2.0
fastapi==0.115.12
fastapi_cors==0.0.6
filelock==3.18.0
flake8==7.1.1
flower==2.0.1
frozenlist==1.6.0
fsspec==2025.5.0
ghp-import==2.1.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
greenlet==3.2.2
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.32.0
humanize==4.12.3
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==5.13.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
limits==5.2.0
litellm==1.70.4
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.5.48
mkdocs-material-extensions==1.3.1
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.2.6
oauthlib==3.2.2
openai==1.82.0
orjson==3.10.0
packaging==25.0
paginate==0.5.7
passlib==1.7.4
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.8
pluggy==1.6.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.7.0
pydantic_core==2.33.2
pyflakes==3.2.0
Pygments==2.19.1
pymdown-extensions==10.15
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.19
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
pyyaml_env_tag==1.1
redis==5.1.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.25.1
rsa==4.9.1
s3transfer==0.10.4
safetensors==0.5.3
shortuuid==1.0.13
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
structlog==24.4.0
text-unidecode==1.3
tiktoken==0.9.0
tokenizers==0.21.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.47.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
vine==5.1.0
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==13.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
